requirements.txt
langchain==0.3.23
langgraph==0.3.34
openai==1.75
tqdm==4.66.2
requests==2.31.0
python-dotenv==1.0.1
